move Navbar to folder organismss
